+++ gallery-link-to-media.php
@@ -24,6 +24,8 @@ define( 'GLTM_PLUGIN_FILE', __FILE__ );
 define( 'GLTM_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
 define( 'GLTM_PLUGIN_URL', plugin_dir_url( __FILE__ ) );
 
+require_once GLTM_PLUGIN_DIR . 'includes/admin.php';
+
 /**
  * Load the Gallery editor behavior.
  */