+++ uninstall.php
@@ -17,3 +17,8 @@ if ( is_multisite() ) {
 	delete_site_option( 'gltm_version' );
 }
 
+delete_post_meta_by_key( '_gltm_original_post_content' );
+delete_post_meta_by_key( '_gltm_original_post_content_date' );
+delete_post_meta_by_key( '_gltm_converted_at' );
+delete_post_meta_by_key( '_gltm_conversion_report' );
+